test: testDelete criado e melhorias no delete controller e service

+++ src/main/java/com/desafio/todo/service/TodoService.java
@@ -46,6 +46,9 @@ public class TodoService {
     }
 
     public List<Todo> delete(UUID ID) {
+        if (!todoRepository.existsById(ID)) {
+            return null;
+        }
         todoRepository.deleteById(ID);
         return list();
     }
